// frontend/src/sections/settings/support/installedPlugins.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InstalledPluginsVariables = Record<string, never>;

export type InstalledPluginsResult = {
  centralServer: {
  plugin: {
  installedPlugins: {
  totalCount: number;
  nodes: Array<{
  id: string;
  code: string;
  version: string;
  kind: "BACKEND" | "FRONTEND";
}>;
};
};
};
};

export const InstalledPlugins = {
  query: "query installedPlugins {\n  centralServer {\n    plugin {\n      installedPlugins {\n        totalCount\n        nodes {\n          id\n          code\n          version\n          kind\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<InstalledPluginsResult, InstalledPluginsVariables>;